+++ tests/StringBytesTest.php
@@ -8,12 +8,7 @@ declare(strict_types=1);
 
 namespace EcomDev\Bytes;
 
-use function memory_get_usage;
-use OutOfBoundsException;
-use OutOfRangeException;
 use PHPUnit\Framework\TestCase;
-use function random_bytes;
-use function str_repeat;
 
 /**
  * Bytes test
@@ -33,39 +28,39 @@ class StringBytesTest extends TestCase
     /** @test */
     public function emptyBytesForEmptyString()
     {
-        $this->assertIterable([], $this->factory->fromString('')->bytes());
+        $this->assertIterable([], $this->factory->createBytesFromString('')->bytes());
     }
 
     /** @test */
     public function zeroSizeForEmptyString()
     {
-        $this->assertEquals(0, $this->factory->fromString('')->size());
+        $this->assertEquals(0, $this->factory->createBytesFromString('')->size());
     }
 
     /** @test */
     public function iteratesOverByteStreamForSingleACharacter()
     {
-        $this->assertIterable(['a'], $this->factory->fromString('a')->bytes());
+        $this->assertIterable(['a'], $this->factory->createBytesFromString('a')->bytes());
     }
 
     /** @test */
     public function returnsOriginalStringSize()
     {
-        $this->assertEquals(5, $this->factory->fromString('abcde')->size());
+        $this->assertEquals(5, $this->factory->createBytesFromString('abcde')->size());
     }
 
     /** @test */
     public function iteratesOverEveryByteInTheString()
     {
-        $this->assertIterable(['a', 'b', 'c', 'd', 'e', 'f', 'g'], $this->factory->fromString('abcdefg')->bytes());
+        $this->assertIterable(['a', 'b', 'c', 'd', 'e', 'f', 'g'], $this->factory->createBytesFromString('abcdefg')->bytes());
     }
 
     /** @test */
     public function sliceFromZeroEqualsTheSameString()
     {
         $this->assertEquals(
-            $this->factory->fromString('i am a string slice'),
-            $this->factory->fromString('i am a string slice')->sliceFrom(0)
+            $this->factory->createBytesFromString('i am a string slice'),
+            $this->factory->createBytesFromString('i am a string slice')->sliceFrom(0)
         );
     }
 
@@ -74,7 +69,7 @@ class StringBytesTest extends TestCase
     {
         $this->assertBytes(
             'slice',
-            $this->factory->fromString('i am a string slice')
+            $this->factory->createBytesFromString('i am a string slice')
                 ->sliceFrom(14)
         );
     }
@@ -84,7 +79,7 @@ class StringBytesTest extends TestCase
     {
         $this->assertEquals(
             5,
-            $this->factory->fromString('i am a slice')->sliceFrom(7)->size()
+            $this->factory->createBytesFromString('i am a slice')->sliceFrom(7)->size()
         );
     }
 
@@ -93,7 +88,7 @@ class StringBytesTest extends TestCase
     {
         $this->assertBytes(
             'second part',
-            $this->factory->fromString('i am a first part second part')
+            $this->factory->createBytesFromString('i am a first part second part')
                 ->sliceFrom(7)
                 ->sliceFrom(11)
         );
@@ -106,7 +101,7 @@ class StringBytesTest extends TestCase
         $this->expectException(OutOfRangeException::class);
         $this->expectExceptionMessage('Bytes are shorter then requested slice');
 
-        $this->factory->fromString('I am short string')
+        $this->factory->createBytesFromString('I am short string')
             ->sliceFrom(10)
             ->sliceFrom(8);
     }
@@ -115,8 +110,8 @@ class StringBytesTest extends TestCase
     public function sliceToTheSameLengthResultsInTheSameSliceAsOriginal()
     {
         $this->assertEquals(
-            $this->factory->fromString('Original slice'),
-            $this->factory->fromString('Original slice')
+            $this->factory->createBytesFromString('Original slice'),
+            $this->factory->createBytesFromString('Original slice')
                 ->sliceTo(14)
         );
     }
@@ -126,7 +121,7 @@ class StringBytesTest extends TestCase
     {
         $this->assertBytes(
             'Only this visible',
-            $this->factory->fromString('Only this visible, not the rest')->sliceTo(17)
+            $this->factory->createBytesFromString('Only this visible, not the rest')->sliceTo(17)
         );
     }
     
@@ -135,7 +130,7 @@ class StringBytesTest extends TestCase
     {
         $this->assertBytes(
             'Middle part.',
-            $this->factory->fromString('Not visible. Middle part. Not visible')
+            $this->factory->createBytesFromString('Not visible. Middle part. Not visible')
                 ->sliceFrom(13)
                 ->sliceTo(12)
         );
@@ -147,7 +142,7 @@ class StringBytesTest extends TestCase
         $this->expectException(OutOfRangeException::class);
         $this->expectExceptionMessage('Bytes are shorter then requested slice');
 
-        $this->factory->fromString('I am short string')
+        $this->factory->createBytesFromString('I am short string')
             ->sliceFrom(10)
             ->sliceTo(8);
     }
@@ -157,7 +152,7 @@ class StringBytesTest extends TestCase
     {
         $this->assertBytes(
             'in the middle',
-            $this->factory->fromString('The most interesting, in the middle, of the text')
+            $this->factory->createBytesFromString('The most interesting, in the middle, of the text')
                 ->sliceTo(35)
                 ->sliceFrom(22)
         );
@@ -169,7 +164,7 @@ class StringBytesTest extends TestCase
         $this->expectException(OutOfBoundsException::class);
         $this->expectExceptionMessage('Slice can only be in positive range');
 
-        $this->factory->fromString('I am short string')
+        $this->factory->createBytesFromString('I am short string')
             ->sliceFrom(-10);
     }
 
@@ -179,7 +174,7 @@ class StringBytesTest extends TestCase
         $this->expectException(OutOfBoundsException::class);
         $this->expectExceptionMessage('Slice can only be in positive range');
 
-        $this->factory->fromString('I am short string')
+        $this->factory->createBytesFromString('I am short string')
             ->sliceTo(-10);
     }
     
@@ -188,7 +183,7 @@ class StringBytesTest extends TestCase
     {
         $this->assertEquals(
             13,
-            $this->factory->fromString('First part, in the middle, second part')
+            $this->factory->createBytesFromString('First part, in the middle, second part')
                 ->sliceTo(25)
                 ->sliceFrom(12)
                 ->size()
@@ -200,13 +195,76 @@ class StringBytesTest extends TestCase
     {
         $this->assertEquals(
             'I am a string.',
-            $this->factory->fromString('Some noise at start. I am a string. Noise in the end.')
+            $this->factory->createBytesFromString('Some noise at start. I am a string. Noise in the end.')
                 ->sliceFrom(21)
                 ->sliceTo(14)
                 ->asString()
         );
     }
 
+    /** @test */
+    public function allowsToPeekIntoFirstCharacterOfAString()
+    {
+        $this->assertEquals(
+            'a',
+            $this->factory->createBytesFromString('abcde')
+                ->peek(0)
+        );
+    }
+    
+    /** @test */
+    public function allowsToPeekIntoTheEndOfTheString()
+    {
+        $this->assertEquals(
+            'e',
+            $this->factory->createBytesFromString('abcde')
+                ->peek(4)
+        );
+    }
+
+    /** @test */
+    public function negativeOffsetErrorsDuringPeek()
+    {
+        $this->expectException(OutOfBoundsException::class);
+        $this->expectExceptionMessage('Position should be always a positive integer');
+
+        $this->factory->createBytesFromString('some-data')
+            ->peek(-1);
+    }
+
+    /** @test */
+    public function outOfBoundPeekOperationTriggersAnError()
+    {
+        $this->expectException(OutOfRangeException::class);
+        $this->expectExceptionMessage('Requested position is not available in the bytes');
+
+        $this->factory->createBytesFromString('Data of 11b')->peek(11);
+    }
+    
+    /** @test */
+    public function slicedFromBytesArePeekedRelatively()
+    {
+        $this->assertEquals(
+            'A',
+            $this->factory
+                ->createBytesFromString('I am not picked slice A I am rest of the slice')
+                ->sliceFrom(20)
+                ->peek(2)
+        );
+    }
+
+    /** @test */
+    public function slicedToLimitsPickedBytesAfterLastCharacter()
+    {
+        $slice = $this->factory->createBytesFromString('I am short slice')->sliceTo(10);
+        $slice->peek(9);
+
+        $this->expectException(OutOfRangeException::class);
+        $this->expectExceptionMessage('Requested position is not available in the bytes');
+
+        $slice->peek(10);
+    }
+    
     /**
      * Asserts iterable result
      */